Add Google Cloud Pub/Sub and JSON functionalities
A 'google-cloud-pubsub' module was added to the requirements.txt file to provide Google Cloud Pub/Sub functionality. In the app.py file, an import statement was included for the 'json' and 'google.cloud.pubsub_v1' modules. This was followed by creating a new function, 'publish', which enables publishing of messages to a Pub/Sub topic, implementing error handling for failed publishing attempts. Also, an environment variable 'TOPIC' was declared to store the Pub/Sub topic name.